# Optimize Lexer comment reading